backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pydantic
numpy
Pillow